# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YGG API",
    "description": "API de análisis de partidas de League of Legends",
    "version": "2.0.0"
  },
  "paths": {
...
    "/matches/player/{player_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List Live Matches",
        "description": "Últimas N partidas del jugador (caché de 1 h), sin filtro de rol ni dependencia de snapshots.",
        "operationId": "matches_list_live_matches",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "live",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Force full fetch from Riot API",
              "default": false,
              "title": "Live"
            },
            "description": "Force full fetch from Riot API"
          },
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Fetch new matches from Riot until stored history is reached",
              "default": false,
              "title": "Sync"
            },
            "description": "Fetch new matches from Riot until stored history is reached"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResponse"
                  },
                  "title": "Response Matches List Live Matches"
                }
              }
            }
          },
          "503": {
            "description": "Riot API unavailable and nothing cached"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/matches/snapshot/{snapshot_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List Matches",
        "description": "Partidas de un snapshot, de la más reciente a la más antigua (solo BD).",
        "operationId": "matches_list_matches",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 50,
              "title": "Limit"
            },
            "description": "Elementos por página"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Elementos a saltar",
              "default": 0,
              "title": "Offset"
            },
            "description": "Elementos a saltar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/matches/snapshot/{snapshot_id}/stats": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Stats",
        "description": "Estadísticas agregadas de un snapshot, calculadas en SQL (queries/snapshot_summary.sql).",
        "operationId": "matches_get_stats",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/matches/player/{player_id}/most-played": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Most Played Champions",
        "description": "Top 3 campeones en las últimas partidas del jugador (desde el historial en BD).",
        "operationId": "matches_get_most_played_champions",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MostPlayedChampionResponse"
                  },
                  "title": "Response Matches Get Most Played Champions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ddragon/items/{item_id}": {
      "get": {
        "tags": [
          "ddragon"
        ],
        "summary": "Get Item",
        "operationId": "ddragon_get_item",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ddragon Get Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ddragon/items/{item_id}/icon": {
      "get": {
        "tags": [
          "ddragon"
        ],
        "summary": "Get Item Icon",
        "operationId": "ddragon_get_item_icon",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ddragon/spells/{key}": {
      "get": {
        "tags": [
          "ddragon"
        ],
        "summary": "Get Spell",
        "operationId": "ddragon_get_spell",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ddragon Get Spell"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ddragon/runes/{rune_id}": {
      "get": {
        "tags": [
          "ddragon"
        ],
        "summary": "Get Rune",
        "operationId": "ddragon_get_rune",
        "parameters": [
          {
            "name": "rune_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rune Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ddragon Get Rune"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ddragon/runes/{rune_id}/icon": {
      "get": {
        "tags": [
          "ddragon"
        ],
        "summary": "Get Rune Icon",
        "operationId": "ddragon_get_rune_icon",
        "parameters": [
          {
            "name": "rune_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rune Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MostPlayedChampionResponse": {
        "properties": {
          "champion_name": {
            "type": "string",
            "title": "Champion Name"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "win_rate": {
            "type": "number",
            "title": "Win Rate"
          },
          "icon_url": {
            "type": "string",
            "title": "Icon Url"
          }
        },
        "type": "object",
        "required": [
          "champion_name",
          "games_played",
          "win_rate",
          "icon_url"
        ],
        "title": "MostPlayedChampionResponse"
      },
...
      "Page_MatchResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MatchResponse]"
      },
...
      "SnapshotStatsResponse": {
        "properties": {
          "snapshot_id": {
            "type": "integer",
            "title": "Snapshot Id"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "winrate": {
            "type": "number",
            "title": "Winrate"
          },
          "avg_kda": {
            "type": "number",
            "title": "Avg Kda"
          },
          "avg_cs_per_min": {
            "type": "number",
            "title": "Avg Cs Per Min"
          },
          "avg_damage_per_min": {
            "type": "number",
            "title": "Avg Damage Per Min"
          },
          "avg_gold_per_min": {
            "type": "number",
            "title": "Avg Gold Per Min"
          },
          "avg_vision": {
            "type": "number",
            "title": "Avg Vision"
          },
          "avg_kill_participation": {
            "type": "number",
            "title": "Avg Kill Participation"
          },
          "first_dragon_rate": {
            "type": "string",
            "title": "First Dragon Rate"
          },
          "herald_rate": {
            "type": "string",
            "title": "Herald Rate"
          },
          "two_or_more_void_grubs_rate": {
            "type": "string",
            "title": "Two Or More Void Grubs Rate"
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "games_played",
          "winrate",
          "avg_kda",
          "avg_cs_per_min",
          "avg_damage_per_min",
          "avg_gold_per_min",
          "avg_vision",
          "avg_kill_participation",
          "first_dragon_rate",
          "herald_rate",
          "two_or_more_void_grubs_rate"
        ],
        "title": "SnapshotStatsResponse"
      },
...
      "UrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "UrlResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}